# docs/swagger.yaml
definitions:
  smart-wardrobe-be_internal_modules_community_application_dto.AddCommentReq:
    properties:
      content:
        type: string
    required:
    - content
    type: object
  smart-wardrobe-be_internal_modules_community_application_dto.CommentRes:
    properties:
      content:
        type: string
      createdAt:
        type: string
      id:
        type: string
      userId:
        type: string
    type: object
  smart-wardrobe-be_internal_modules_community_application_dto.CreatePostReq:
    properties:
      contactInfo:
        type: string
      content:
        type: string
      itemIds:
        items:
          type: string
        type: array
      media:
        items:
          $ref: '#/definitions/smart-wardrobe-be_internal_modules_community_application_dto.PostMediaReq'
        type: array
      postType:
        type: string
      title:
        type: string
      totalPrice:
        type: number
    required:
    - content
    - postType
    type: object
  smart-wardrobe-be_internal_modules_community_application_dto.LikePostReq:
    properties:
      isLiked:
        type: boolean
    required:
    - isLiked
    type: object
  smart-wardrobe-be_internal_modules_community_application_dto.PostItemRes:
    properties:
      buyerUserId:
        type: string
      id:
        type: string
      item:
        $ref: '#/definitions/smart-wardrobe-be_internal_modules_wardrobe_application_dto.WardrobeItemRes'
      itemCondition:
        type: integer
      price:
        type: number
      soldAt:
        type: string
      status:
        type: integer
      transferState:
        type: integer
    type: object
  smart-wardrobe-be_internal_modules_community_application_dto.PostMediaReq:
    properties:
      mediaType:
        type: string
      mediaUrl:
        type: string
      publicId:
        type: string
      sortOrder:
        type: integer
    required:
    - mediaType
    - mediaUrl
    type: object
  smart-wardrobe-be_internal_modules_community_application_dto.PostMediaRes:
    properties:
      id:
        type: string
      mediaType:
        type: string
      mediaUrl:
        type: string
      publicId:
        type: string
      sortOrder:
        type: integer
    type: object
  smart-wardrobe-be_internal_modules_community_application_dto.PostRes:
    properties:
      commentCount:
        type: integer
      comments:
        items:
          $ref: '#/definitions/smart-wardrobe-be_internal_modules_community_application_dto.CommentRes'
        type: array
      contactInfo:
        type: string
      content:
        type: string
      createdAt:
        type: string
      id:
        type: string
      items:
        items:
          $ref: '#/definitions/smart-wardrobe-be_internal_modules_community_application_dto.PostItemRes'
        type: array
      likeCount:
        type: integer
      media:
        items:
          $ref: '#/definitions/smart-wardrobe-be_internal_modules_community_application_dto.PostMediaRes'
        type: array
      postType:
        type: string
      title:
        type: string
      totalPrice:
        type: number
      updatedAt:
        type: string
      userId:
        type: string
    type: object
  smart-wardrobe-be_internal_modules_community_application_dto.RemovePostItemsReq:
    properties:
      postItemIds:
        items:
          type: string
        type: array
    required:
    - postItemIds
    type: object
  smart-wardrobe-be_internal_modules_community_application_dto.UpdatePostItemsBuyerReq:
    properties:
      buyerUserId:
        type: string
    required:
    - buyerUserId
    type: object
...
  smart-wardrobe-be_internal_modules_wardrobe_application_dto.CategoryRes:
    properties:
      id:
        type: string
      name:
        type: string
      slug:
        type: string
    type: object
...
  smart-wardrobe-be_internal_modules_wardrobe_application_dto.WardrobeItemRes:
    properties:
      category:
        $ref: '#/definitions/smart-wardrobe-be_internal_modules_wardrobe_application_dto.CategoryRes'
      color:
        type: string
      createdAt:
        type: string
      fit:
        type: string
      id:
        type: string
      imagePublicId:
        type: string
      imageUrl:
        type: string
      isLocked:
        type: boolean
      material:
        type: string
      pattern:
        type: string
      seasonality:
        type: string
      status:
        $ref: '#/definitions/smart-wardrobe-be_internal_shared_domain_constants_wardrobestatus.WardrobeItemStatus'
      style:
        type: string
      userId:
        type: string
    type: object
...
  smart-wardrobe-be_internal_shared_domain_constants_wardrobestatus.WardrobeItemStatus:
    enum:
    - 0
    - 1
    - 2
    - 3
    - 4
    format: int32
    type: integer
    x-enum-comments:
      Failed: AI processing failed
      Processing: AI processing in background
    x-enum-descriptions:
    - ""
    - ""
    - ""
    - AI processing in background
    - AI processing failed
    x-enum-varnames:
    - InWardrobe
    - Selling
    - Sold
    - Processing
    - Failed
  smart-wardrobe-be_internal_shared_presentation.APIResponse:
    properties:
      data: {}
      message:
        type: string
    type: object
info:
  contact: {}
  description: API docs for Smart Wardrobe.
  title: Smart Wardrobe API Document
  version: "1.0"
paths:
...
  /api/v1/posts:
    get:
      description: Lấy feed danh sách bài đăng của cộng đồng sắp xếp theo thứ tự mới
        nhất/hot nhất
      produces:
      - application/json
      responses:
        "200":
          description: Lấy feed thành công
          schema:
            allOf:
            - $ref: '#/definitions/smart-wardrobe-be_internal_shared_presentation.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/smart-wardrobe-be_internal_modules_community_application_dto.PostRes'
                  type: array
              type: object
      summary: Lấy danh sách bài đăng cộng đồng
      tags:
      - Community
    post:
      consumes:
      - application/json
      description: Đăng bài bán đồ hoặc khoe outfit lên bảng tin cộng đồng
      parameters:
      - description: Nội dung bài đăng
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/smart-wardrobe-be_internal_modules_community_application_dto.CreatePostReq'
      produces:
      - application/json
      responses:
        "201":
          description: Tạo bài đăng thành công
          schema:
            allOf:
            - $ref: '#/definitions/smart-wardrobe-be_internal_shared_presentation.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/smart-wardrobe-be_internal_modules_community_application_dto.PostRes'
              type: object
      summary: Tạo bài đăng cộng đồng mới
      tags:
      - Community
  /api/v1/posts/{postID}:
    delete:
      description: Xóa bài đăng của chính người dùng hiện tại
      parameters:
      - description: ID bài đăng
        in: path
        name: postID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Xóa bài đăng thành công
          schema:
            $ref: '#/definitions/smart-wardrobe-be_internal_shared_presentation.APIResponse'
      summary: Xóa bài đăng
      tags:
      - Community
    get:
      description: Lấy thông tin chi tiết của một bài đăng cụ thể kèm danh sách bình
        luận
      parameters:
      - description: ID bài đăng
        in: path
        name: postID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Lấy chi tiết bài đăng thành công
          schema:
            allOf:
            - $ref: '#/definitions/smart-wardrobe-be_internal_shared_presentation.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/smart-wardrobe-be_internal_modules_community_application_dto.PostRes'
              type: object
      summary: Lấy chi tiết bài đăng
      tags:
      - Community
  /api/v1/posts/{postID}/comments:
    post:
      consumes:
      - application/json
      description: Tạo bình luận mới dưới bài viết cộng đồng
      parameters:
      - description: ID bài đăng
        in: path
        name: postID
        required: true
        type: string
      - description: Nội dung bình luận
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/smart-wardrobe-be_internal_modules_community_application_dto.AddCommentReq'
      produces:
      - application/json
      responses:
        "201":
          description: Thêm bình luận thành công
          schema:
            allOf:
            - $ref: '#/definitions/smart-wardrobe-be_internal_shared_presentation.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/smart-wardrobe-be_internal_modules_community_application_dto.CommentRes'
              type: object
      summary: Thêm bình luận vào bài viết
      tags:
      - Community
  /api/v1/posts/{postID}/items:
    delete:
      consumes:
      - application/json
      description: Gỡ một hoặc nhiều món đồ ra khỏi danh sách bán trong bài đăng
      parameters:
      - description: ID bài đăng
        in: path
        name: postID
        required: true
        type: string
      - description: Danh sách ID các món đồ cần gỡ
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/smart-wardrobe-be_internal_modules_community_application_dto.RemovePostItemsReq'
      produces:
      - application/json
      responses:
        "200":
          description: Gỡ món đồ thành công
          schema:
            $ref: '#/definitions/smart-wardrobe-be_internal_shared_presentation.APIResponse'
      summary: Gỡ món đồ khỏi bài đăng
      tags:
      - Community
  /api/v1/posts/{postID}/like:
    put:
      consumes:
      - application/json
      description: Like hoặc unlike một bài viết trên cộng đồng bằng cách gửi trạng
        thái rõ ràng
      parameters:
      - description: ID bài đăng
        in: path
        name: postID
        required: true
        type: string
      - description: Trạng thái thích
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/smart-wardrobe-be_internal_modules_community_application_dto.LikePostReq'
      produces:
      - application/json
      responses:
        "200":
          description: Cập nhật like thành công
          schema:
            $ref: '#/definitions/smart-wardrobe-be_internal_shared_presentation.APIResponse'
      summary: Thích / Bỏ thích bài đăng
      tags:
      - Community
...
  /api/v1/transfers/{postItemID}/accept:
    post:
      description: Đồng ý nhận trang phục đã mua về tủ đồ cá nhân (Trang phục sẽ chuyển
        quyền sở hữu)
      parameters:
      - description: ID chi tiết món đồ trong bài post
        in: path
        name: postItemID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Nhận món đồ vào tủ thành công
          schema:
            allOf:
            - $ref: '#/definitions/smart-wardrobe-be_internal_shared_presentation.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/smart-wardrobe-be_internal_modules_community_application_dto.PostItemRes'
              type: object
      summary: Chấp nhận nhận bàn giao trang phục
      tags:
      - Community
  /api/v1/transfers/{postItemID}/decline:
    post:
      description: Từ chối nhận bàn giao trang phục mua từ bài đăng cộng đồng
      parameters:
      - description: ID chi tiết món đồ trong bài post
        in: path
        name: postItemID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Từ chối nhận món đồ thành công
          schema:
            $ref: '#/definitions/smart-wardrobe-be_internal_shared_presentation.APIResponse'
      summary: Từ chối nhận bàn giao trang phục
      tags:
      - Community
  /api/v1/transfers/{postItemID}/mark-sold:
    post:
      consumes:
      - application/json
      description: Đánh dấu trang phục đã được bán cho một người dùng khác và kích
        hoạt trạng thái bàn giao (transfer)
      parameters:
      - description: ID chi tiết món đồ trong bài post
        in: path
        name: postItemID
        required: true
        type: string
      - description: Thông tin người mua
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/smart-wardrobe-be_internal_modules_community_application_dto.UpdatePostItemsBuyerReq'
      produces:
      - application/json
      responses:
        "200":
          description: Đánh dấu đã bán thành công
          schema:
            $ref: '#/definitions/smart-wardrobe-be_internal_shared_presentation.APIResponse'
      summary: Đánh dấu món đồ đã bán
      tags:
      - Community
...
swagger: "2.0"